Job name translation
