+++ packages/core/src/datamanager_new.ts
@@ -27,89 +27,16 @@ import {
 import {BaseParamToStatKey, RelevantBaseParam} from "./external/xivapitypes";
 import {getRelicStatModelFor} from "./relicstats/relicstats";
 import {requireNumber, requireString} from "./external/data_validators";
-import {
-    DataApiClient,
-    FoodStatBonus,
-    GearAcquisitionSource as AcqSrc,
-    HttpResponse
-} from "@xivgear/data-api-client/dataapi";
+import {DataApiClient, FoodStatBonus, GearAcquisitionSource as AcqSrc} from "@xivgear/data-api-client/dataapi";
 import {BaseParamMap, DataManager, DmJobs} from "./datamanager";
 import {IlvlSyncInfo} from "./datamanager_xivapi";
 import {applyStatCaps} from "./gear";
 import {toTranslatable, TranslatableString} from "@xivgear/i18n/translation";
 import {RawStatsPart} from "@xivgear/util/types";
+import {API_CLIENT, ApiFoodData, ApiItemData, ApiMateriaData, checkResponse} from "./data_api_client";
+
 // import {recordError} from "@xivgear/common-ui/analytics/analytics";
 
-type ApiClientRawType<X extends keyof DataApiClient<never>, Y extends keyof DataApiClient<never>[X]> = DataApiClient<never>[X][Y]
-
-type ApiItemData = Awaited<ReturnType<ApiClientRawType<'items', 'items'>>>['data']['items'][number]
-type ApiMateriaData = Awaited<ReturnType<ApiClientRawType<'materia', 'materia'>>>['data']['items'][number]
-type ApiFoodData = Awaited<ReturnType<ApiClientRawType<'food', 'foodItems'>>>['data']['items'][number]
-// type BaseParamType = Awaited<ReturnType<ApiClientRawType<'baseParams', 'baseParams'>>>['data']['items'][number]
-// type JobType = Awaited<ReturnType<ApiClientRawType<'jobs', 'jobs'>>>['data']['items'][number]
-// type ItemLevelType = Awaited<ReturnType<ApiClientRawType<'itemLevel', 'itemLevels'>>>['data']['items'][number]
-
-export type DataManagerErrorReporter = (r: Response, params: Parameters<typeof fetch>) => void;
-
-let errorReporter: DataManagerErrorReporter = () => {
-};
-
-export function setDataManagerErrorReporter(reporter: DataManagerErrorReporter) {
-    errorReporter = reporter;
-}
-
-async function retryFetch(...params: Parameters<typeof fetch>): Promise<Response> {
-    let tries = 5;
-    while (true) {
-        tries--;
-        const result = await fetch(...params);
-        // TODO: add other errors here?
-        if (tries > 0 && result.status === 503 && result.statusText === 'Not Ready') {
-            console.log("data api not ready, retrying", params[0]);
-            await new Promise(r => setTimeout(r, 5_000 + (Math.random() * 1000)));
-            continue;
-        }
-        if (result.status >= 400) {
-            const content = JSON.stringify(await result.json());
-            console.error(`Data API error: ${result.status}: ${result.statusText}`, params[0], content);
-            const error = Error(`Data API error: ${result.status}: ${result.statusText} (${params[0]}\n${content}`);
-            // TODO: it would be nice to be able to record these, but they would create an unacceptable
-            // dependency loop.
-            // recordError("datamanager", error, {fetchUrl: String(params[0])});
-            throw error;
-        }
-        if (!result.ok) {
-            console.error(`Data API error: ${result.status}: ${result.statusText}`, params[0]);
-            errorReporter(result, params);
-            throw new Error(`Data API error: ${result.status}: ${result.statusText} (${params[0]}`);
-        }
-        return result;
-    }
-}
-
-const apiClient = new DataApiClient<never>({
-    baseUrl: "https://data.xivgear.app",
-    // baseUrl: "https://betadata.xivgear.app",
-    // baseUrl: "http://localhost:8085",
-    customFetch: retryFetch,
-});
-
-export function setDataApi(baseUrl: string) {
-    apiClient.baseUrl = baseUrl;
-}
-
-function checkResponse<X extends HttpResponse<unknown>>(response: X): X {
-    const url = response.url;
-    if (!response.ok) {
-        errorReporter(response, [url]);
-        throw new Error(`Data API error: response not ok: ${response.status}: ${response.statusText} (${url}`);
-    }
-    else if (!response.data) {
-        errorReporter(response, [url]);
-        throw new Error(`Data API error: no data: ${response.status}: ${response.statusText} (${url}`);
-    }
-    return response;
-}
 
 export class NewApiDataManager implements DataManager {
 
@@ -133,7 +60,7 @@ export class NewApiDataManager implements DataManager {
         this._minIlvlFood = lvlData.minILvlFood;
         this._maxIlvlFood = lvlData.maxILvlFood;
         this._ilvlSync = ilvlSync;
-        this.apiClient = apiClient;
+        this.apiClient = API_CLIENT;
     }
 
     private _allItems: DataApiGearInfo[];
